move footer into sidebar
